+++ candidate/src/main/java/com/lawencon/candidate/dto/candidate/CandidatePasswordUpdateReqDto.java
@@ -1,0 +1,24 @@
+package com.lawencon.candidate.dto.candidate;
+
+public class CandidatePasswordUpdateReqDto {
+
+	private String oldPassword;
+	private String newPassword;
+
+	public String getOldPassword() {
+		return oldPassword;
+	}
+
+	public void setOldPassword(String oldPassword) {
+		this.oldPassword = oldPassword;
+	}
+
+	public String getNewPassword() {
+		return newPassword;
+	}
+
+	public void setNewPassword(String newPassword) {
+		this.newPassword = newPassword;
+	}
+
+}
