+++ candidate/src/main/java/com/lawencon/candidate/service/CandidateProfileService.java
@@ -1,0 +1,10 @@
+package com.lawencon.candidate.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CandidateProfileService {
+	
+	
+
+}
